sanphelp_microservice/mobile_notifications/requirements.txt
amqp==1.4.8
anyjson==0.3.3
astroid==1.4.6
billiard==3.3.0.22
celery==3.1.19
Django==1.8.4
django-celery==3.1.17
django-push-notifications==1.4.1
djangorestframework==3.2.4
django-rest-swagger==0.3.5
enum34==1.0.4
MySQL-python==1.2.5
